Mitigates many issues with the regular elevator algorithm
Regular elevator algorithm is much closer to completion
There is a small problem where sometimes elevators will stop accepting requests. I believe this has to do with another elevator "stealing" jobs, while the elevator is currently opening its door to accept said job, if this happens to the cab during the trip it will return to ground floor but never accept any more requests. 
There are a few performance changes that need to be made in order to meet 8x speed