Regular Elevator Algorithm Finished
There are some minor issues, but the algorithm never crashes or produces a wrong result. Elevators will continue to pick up passengers, and won't stop until the building is fully satisfied. 
Issue 1: sometimes, (believed to be with a task getting distributed "stolen" multiple times for elevators > 3) an elevator will open its door innappropriately on a previous floor that it thought to have a request at
Issue 2: for n > 4 elevators and simulation speed > 1, if there is a lot of animation with doors openingclosing code it seems as though this may potentially affect the speed of the overall program. Frames, and statistics remain accurate, however the actual simulation speed is not maintained when there is a large amount of animations being rendered
There are no other known issues with the Regular Or Covid-19 algorithm